+++ src/lib/helpers/typeahead.tsx
@@ -1,53 +1,144 @@
-import React from "react";
-import { Option } from "react-bootstrap-typeahead/types/types";
-import { Menu, MenuItem } from "react-bootstrap-typeahead";
-import { RenderMenuProps } from "react-bootstrap-typeahead/types/components/Typeahead";
+import { AutocompleteRenderOptionState } from "@mui/material/Autocomplete";
 import { LabelValueOption } from "../types/LabelValueOption";
+import { TypeaheadOption } from "../types/Typeahead";
+import AutosuggestHighlightMatch from "autosuggest-highlight/match";
+import AutosuggestHighlightParse from "autosuggest-highlight/parse";
+import { SxProps } from "@mui/material";
 
-const convertTypeaheadOptionsToStringArray = (options: Option[]): string[] => {
-  const isStringArray = options.length > 0 && options.every((value) => typeof value === "string");
+const isStringArray = (options: TypeaheadOption[]): boolean => options.length > 0 && options.every((value) => typeof value === "string");
 
-  if (isStringArray) {
+const convertAutoCompleteOptionsToStringArray = (options: TypeaheadOption[] | undefined): string[] => {
+  if (!options) {
+    return [];
+  }
+
+  if (isStringArray(options)) {
     return options as string[];
   }
 
-  return (options as Record<string, string>[]).map((option) => option.value);
+  return (options as LabelValueOption[]).map((option) => option.value) as string[];
 };
 
-const renderMenu = (results: LabelValueOption[], menuProps: RenderMenuProps): JSX.Element => {
-  const groups = [...new Set(results.filter((x) => x.group?.name).map((option) => option.group?.name))];
-  const anonymousOptions = results.filter((option) => !option.group?.name);
-  let position = 0;
+const getSingleAutoCompleteValue = (options: TypeaheadOption[], value: string | number | undefined) => {
+  if (options[0] === undefined || value === undefined) {
+    return undefined;
+  }
 
+  return options.find((x) => (typeof x === "string" ? x === value : x.value == value));
+};
+
+const getMultipleAutoCompleteValue = (options: TypeaheadOption[], value: string[] | undefined) => {
+  if (options[0] === undefined || !value) {
+    return undefined;
+  }
+  return options.filter((x) => (typeof x === "string" ? value.includes(x) : value.includes(x.value as string))) as
+    | string[]
+    | LabelValueOption[]
+    | undefined;
+};
+
+const sortOptionsByGroup = (options: TypeaheadOption[]): TypeaheadOption[] =>
+  options.sort((x, y) => (typeof x === "string" ? x : x.group?.name ?? "").localeCompare(typeof y === "string" ? y : y.group?.name ?? ""));
+
+const groupOptions = (option: TypeaheadOption): string => (typeof option === "string" ? option : option.group?.name ?? "");
+
+const isDisabledGroup = (option: TypeaheadOption): boolean => typeof option !== "string" && !!option.group?.disabled;
+
+function getUniqueOptions(source: TypeaheadOption[], comparison: TypeaheadOption[]) {
+  const comparisonSet = new Set(comparison.map((x) => (typeof x === "string" ? x : x.value)));
+  return source.filter((x) => (typeof x === "string" ? !comparisonSet.has(x) : !comparisonSet.has(x.value)));
+}
+
+const renderHighlightedOptionFunction = (
+  props: React.HTMLAttributes<HTMLLIElement>,
+  option: TypeaheadOption,
+  { inputValue }: AutocompleteRenderOptionState,
+) => {
+  const finalOption = typeof option === "string" ? option : option.label;
+  const matches = AutosuggestHighlightMatch(finalOption, inputValue, { insideWords: true });
+  const parts = AutosuggestHighlightParse(finalOption, matches) as Array<{ text: string; highlight: boolean }>;
   return (
-    <Menu {...menuProps}>
-      {groups.map((group, index) => (
-        <React.Fragment key={index}>
-          <Menu.Header>{group}</Menu.Header>
-          {results
-            .filter((x) => x.group?.name === group)
-            .map((option) => (
-              <MenuItem
-                key={option.value}
-                option={option}
-                position={position++}
-                disabled={option.disabled || option.group?.disabled}
-                className="ps-4"
-              >
-                {option.label}
-              </MenuItem>
-            ))}
-          {index < groups.length - 1 && <Menu.Divider />}
-        </React.Fragment>
-      ))}
-      {!!anonymousOptions.length && <Menu.Divider />}
-      {anonymousOptions.map((option) => (
-        <MenuItem key={option.value} option={option} position={position++} disabled={option.disabled}>
-          {option.label}
-        </MenuItem>
-      ))}
-    </Menu>
+    <li {...props}>
+      <div>
+        {parts.map((part, index) => (
+          <span key={index} style={{ fontWeight: part.highlight ? 700 : 400 }}>
+            {part.text}
+          </span>
+        ))}
+      </div>
+    </li>
   );
 };
 
-export { convertTypeaheadOptionsToStringArray, renderMenu };
+const bootstrapStyle: SxProps = {
+  ".MuiOutlinedInput-root": {
+    fontFamily: [
+      "-apple-system",
+      "BlinkMacSystemFont",
+      '"Segoe UI"',
+      "Roboto",
+      '"Helvetica Neue"',
+      "Arial",
+      "sans-serif",
+      '"Apple Color Emoji"',
+      '"Segoe UI Emoji"',
+      '"Segoe UI Symbol"',
+    ].join(","),
+    border: "1px solid #ced4da",
+    fontSize: "1rem",
+    color: "#495057",
+    paddingY: "0.5px !important",
+    paddingLeft: "0.1px !important",
+    borderColor: "#E0E3E7",
+    transition: "border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out",
+    "&.Mui-focused": {
+      borderColor: "#80bdff",
+      boxShadow: "0 0 0 0.2rem rgba(0,123,255,.25)",
+    },
+    "& .MuiAutocomplete-input": {
+      padding: "0.375rem 0.75rem !important",
+    },
+    "&.Mui-error": {
+      borderColor: "#dc3545",
+    },
+    "&.Mui-focused.Mui-error": {
+      boxShadow: "0 0 0 0.2rem rgba(220,53,69,.25)",
+    },
+    "& .MuiInputAdornment-positionStart": {
+      paddingLeft: 1.5,
+    },
+    "& .MuiChip-root": {
+      backgroundColor: "#e9ecef",
+      color: "#495057",
+      fontSize: "0.875rem",
+      borderRadius: "0.25rem",
+      display: "flex",
+    },
+    "& .MuiChip-deleteIcon": {
+      color: "#495057",
+      fontSize: "0.875rem",
+    },
+  },
+  "& fieldset": {
+    border: "none",
+  },
+  "& .MuiInputLabel-root": {
+    marginTop: "-1rem",
+    marginLeft: "-0.8rem",
+  },
+  "& .MuiFormHelperText-root ": {
+    marginLeft: "0.2rem",
+    marginTop: "0.3rem",
+  },
+};
+export {
+  getSingleAutoCompleteValue,
+  getMultipleAutoCompleteValue,
+  convertAutoCompleteOptionsToStringArray,
+  sortOptionsByGroup,
+  isDisabledGroup,
+  groupOptions,
+  getUniqueOptions,
+  renderHighlightedOptionFunction,
+  bootstrapStyle,
+};