+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- :boom: `AsyncTypeAheadInput` based on MUI Autocomplete component.
+- :boom: `StaticTypeAheadInput` based on MUI Autocomplete component.
+
 ## [2.13.1] - 2024-12-02
 
 ### Fixed